test: We have the initial workings of a test suite
This also shifts the storage of supported markets to something done
during runtime. A call is made against the AddressResolver to find the
market. If said market does not exist then the tx is reverted. This
might pose a problem in that any valid response from the AddressResolver
may cause unintended issues even if they do hit subscribe to the
FuturesMarket base. To be resolved later.