feat: Add support for 721 transfers
How does this work?

A `manager` address is passed to the Pozition on clone/init. After, when
an _mint, _burn, or any variation of _transfer is invoked, post
operation, a hook is invoked to update the allMintedPositions in the
manager.

The update understands at this point, the position has been transferred
hence _to is expected to have the address and _from is assumed to have
an existing contract. However, this is only true _iff_ the operation is
_not_ a _mint. The logic is slightly varied for _mint operations as the
_to is addr(0). This is also true for _burn as the _from will be
addr(0).

Addition to this work, now that Pozition has a reference to the manager,
a modifier can be used to only allow certain functions callable via the
manager.